Add PackageDetailsCard and PackageLocationCard
